+++ database/migrations/2020_04_01_202034_create_inventories_table.php
@@ -14,15 +14,15 @@ class CreateInventoriesTable extends Migration
     public function up()
     {
         Schema::create('inventories', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('item_id');
-            $table->unsignedInteger('color_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('item_id');
+            $table->unsignedBigInteger('color_id');
             $table->integer('qty_stock');
-            $table->unsignedInteger('size_id');
+            $table->unsignedBigInteger('size_id');
             $table->integer('sort');
             $table->string('available_on');
             $table->boolean('active')->default(1)->change();
-            $table->unsignedInteger('user_id');
+            $table->unsignedBigInteger('user_id');
             $table->softDeletes();
             $table->timestamps();
