+++ database/migrations/2020_04_02_023648_create_buyer_shipping_addresses_table.php
@@ -14,7 +14,7 @@ class CreateBuyerShippingAddressesTable extends Migration
     public function up()
     {
         Schema::create('buyer_shipping_addresses', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('location');
             $table->text('address');
             $table->string('country');
@@ -24,8 +24,8 @@ class CreateBuyerShippingAddressesTable extends Migration
             $table->integer('phone');
             $table->string('fax');
             $table->boolean('active')->default(1)->change();
-            $table->unsignedInteger('buyer_id');
-            $table->unsignedInteger('user_id');
+            $table->unsignedBigInteger('buyer_id');
+            $table->unsignedBigInteger('user_id');
             $table->softDeletes();
             $table->timestamps();
